% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-addin.R
\name{add_insight_addin}
\alias{add_insight_addin}
\title{Driver function for data.world Insight Add-in}
\usage{
add_insight_addin()
}
\description{
Driver function for data.world Insight Add-in
}
\keyword{internal}
