% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFile.R
\name{addFile}
\alias{addFile}
\title{Add a file to a container}
\usage{
addFile(request, name, url)
}
\arguments{
\item{request}{some object with a file container}

\item{name}{the filename including the file extension. If a file by that name
already exists in the dataset, the file will be updated/overwritten.}

\item{url}{the public, full direct-download path to file}
}
\description{
Add a file to a container
}
\examples{
fileBatchUpdateRequest <- data.world::FileBatchUpdateRequest()

fileBatchUpdateRequest <- data.world::addFile(request = fileBatchUpdateRequest,
   name = "file.csv", url = "https://data.world/file3.csv")

createDatasetRequest <- data.world::DatasetCreateRequest(title="coffeeCounty",
    visibility = "OPEN", description = "coffee county , AL - census income" ,
    tags = c("rsdk", "sdk", "arr") , licenseString = "Public Domain")

createDatasetRequest <- data.world::addFile(request = createDatasetRequest,
    name = "file4.csv", url = "https://data.world/file4.csv")

datasetPutRequest <- data.world::DatasetPutRequest(visibility = "OPEN",
    description = "updated description", files = list())

datasetPutRequest <- data.world::addFile(request = datasetPutRequest,
    name = "file4.csv", url = "https://data.world/file4.csv")
}
