% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFilesBySource.R
\name{addFileBySource.ApiClient}
\alias{addFileBySource.ApiClient}
\title{publish a single file on the web to be added to a data.world dataset}
\usage{
\method{addFileBySource}{ApiClient}(apiClient, dataset, name, url)
}
\arguments{
\item{apiClient}{an apiClient}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{name}{the filename including the file extension. If a file by that name already
exists in the dataset, the file will be updated/overwritten.}

\item{url}{the public, full direct-download path to file}
}
\description{
publish a single file on the web to be added to a data.world dataset
}
\examples{
data.world::addFileBySource(data.world()$apiClient, dataset = "ownerid/datasetid",
    name = "file.csv", url = "https://data.world/some_file.csv")
}
