% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg}
\alias{cfg}
\title{Construct configuration object using runtime values.}
\usage{
cfg(auth_token)
}
\arguments{
\item{auth_token}{API authorization token.}
}
\value{
Object of type \code{cfg}
}
\description{
Construct configuration object using runtime values.
}
\examples{
runtime_cfg <- data.world::cfg(auth_token = "YOUR_TOKEN")
data.world::set_config(runtime_cfg)
}
