% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_env}
\alias{cfg_env}
\title{Construct configuration object using environment variables.}
\usage{
cfg_env(auth_token_var = "DW_AUTH_TOKEN")
}
\arguments{
\item{auth_token_var}{Name of environment variable for API authorization
token (default: \code{DW_AUTH_TOKEN}).}
}
\value{
Object of type \code{cfg_env}
}
\description{
Construct configuration object using environment variables.
}
\examples{
envvars_cfg <- data.world::cfg_env()
data.world::set_config(envvars_cfg)
}
