% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_saved}
\alias{cfg_saved}
\title{Construct configuration object using file-based configuration.}
\usage{
cfg_saved(profile = "DEFAULT")
}
\arguments{
\item{profile}{Name of configuration profile (default: \code{DEFAULT}).}
}
\value{
Object of type \code{cfg_saved}
}
\description{
Construct configuration object using file-based configuration.
}
\examples{
saved_cfg <- data.world::cfg_saved()
data.world::set_config(saved_cfg)
}
