% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetSummaryResponse.R
\name{checkDatasetSummaryResponse}
\alias{checkDatasetSummaryResponse}
\title{Validating the required fields are there}
\usage{
checkDatasetSummaryResponse(object)
}
\arguments{
\item{object}{the (supposedly) DatasetSummaryResponse object}
}
\description{
Validating the required fields are there
}
