% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{configure_package}
\alias{configure_package}
\title{Apply configuration from file or envvars and load dwapi}
\usage{
configure_package()
}
\description{
Apply configuration from file or envvars and load dwapi
}
\keyword{internal}
