% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createDataset.ApiClient}
\alias{createDataset.ApiClient}
\title{create a dataset and add metadata}
\usage{
\method{createDataset}{ApiClient}(apiClient, createDatasetRequest, ownerId)
}
\arguments{
\item{apiClient}{a data.world client}

\item{createDatasetRequest}{the request payload}

\item{ownerId}{Dataset owner username. Included in data.world dataset URL path, after domain.}
}
\description{
create a dataset and add metadata
}
\examples{
request <- data.world::DatasetCreateRequest(title="testdataset", visibility = "OPEN",
     description = "Test Dataset by R-SDK" , tags = c("rsdk", "sdk", "arr") ,
     licenseString = "Public Domain")
request <- data.world::addFile(request = request, name = "file4.csv",
     url = "https://data.world/file4.csv")
data.world::createDataset(apiClient = data.world()$apiClient, createDatasetRequest = request,
     ownerId = "ownerid")
}
