% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createDataset}
\alias{createDataset}
\title{create a dataset and add metadata}
\usage{
createDataset(connection, createDatasetRequest, ownerId)
}
\arguments{
\item{connection}{the connection to data.world}

\item{createDatasetRequest}{the request payload}

\item{ownerId}{Dataset owner username. Included in data.world dataset URL path, after domain.}
}
\description{
create a dataset and add metadata
}
\examples{
request <- data.world::DatasetCreateRequest(title="testdataset", visibility = "OPEN", description = "Test Dataset by R-SDK" , tags = c("rsdk", "sdk", "arr") , licenseString = "Public Domain")
request <- data.world::addFile(request = request, name = "file4.csv", url = "https://data.world/file4.csv")
conn <- data.world(token = "YOUR_API_TOKEN_HERE")
data.world::createDataset(connection = conn, createDatasetRequest = request, ownerId = "ownerid")
}
