% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-addin.R
\name{current_plot_exists}
\alias{current_plot_exists}
\title{Determine if there is a plot in the RStudio Plots view}
\usage{
current_plot_exists()
}
\value{
true if a plot exists
}
\description{
Determine if there is a plot in the RStudio Plots view
}
\keyword{internal}
