% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.world-deprecated.R
\docType{package}
\name{data.world-deprecated}
\alias{data.world-deprecated}
\alias{query.data.world}
\alias{data.world-deprecated-package}
\title{Deprecated function(s) in the data.world package}
\description{
These functions have been marked for removal from the data.world package.
}
\section{Details}{

\tabular{ll}{
   \code{data.world} \tab is replaced by \code{\link{set_config}}\cr
   \code{query.data.world} \tab is replaced by \code{\link{query}}\cr
}
}

