% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{DataPackage}
\alias{DataPackage}
\title{An adapter class to data package structure generated by https://github.com/frictionlessdata/datapackage-r}
\usage{
DataPackage(structure, dataset)
}
\arguments{
\item{structure}{- a structure object returned `datapkg::datapkg_read`}

\item{dataset}{the  "agentid/datasetid" for the dataset from
which to download the datapackage}
}
\description{
An adapter class to data package structure generated by https://github.com/frictionlessdata/datapackage-r
}
\examples{
\dontrun{
cities <- datapkg_read("git://github.com/datasets/world-cities")
package <- data.world::dataPackage(cities)
}
}
