% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetCreateRequest.R
\name{DatasetCreateRequest}
\alias{DatasetCreateRequest}
\title{DatasetCreateRequest per https://api.data.world/v0/swagger.json}
\usage{
DatasetCreateRequest(title, visibility, description = "", summary = "",
  tags = list(), licenseString = "", fileCreateRequests = list())
}
\arguments{
\item{title}{Dataset name. 3 to 30 characters (required)}

\item{visibility}{PRIVATE or OPEN (required)}

\item{description}{Short description for dataset.}

\item{summary}{Dataset summary, markdown supported}

\item{tags}{Array of strings to tag datase}

\item{licenseString}{Public Domain, PDDL, CC-0, CC-BY, ODC-BY, CC-BY-SA, ODC-ODbL, CC BY-NC-SA}

\item{fileCreateRequests}{list of `FileCreateRequest`}
}
\description{
DatasetCreateRequest per https://api.data.world/v0/swagger.json
}
\examples{
request <- data.world::DatasetCreateRequest(title="coffeeCounty", visibility = "OPEN", description = "coffee county , AL - census income" , tags = c("rsdk", "sdk", "arr") , licenseString = "Public Domain")
request <- data.world::addFile(request = request, name = "file.csv", url = "http://data.world/file.csv")
}
