% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetPatchRequest.R
\name{DatasetPatchRequest}
\alias{DatasetPatchRequest}
\title{DatasetPatchRequest per https://api.data.world/v0/swagger.json}
\usage{
DatasetPatchRequest(description = NULL, summary = NULL, tags = NULL,
  license = NULL, visibility = NULL, files = NULL)
}
\arguments{
\item{description}{Short description for dataset.}

\item{summary}{Dataset summary, markdown supported}

\item{tags}{Array of strings to tag datase}

\item{license}{Public Domain, PDDL, CC-0, CC-BY, ODC-BY, CC-BY-SA, ODC-ODbL, CC BY-NC-SA}

\item{visibility}{PRIVATE or OPEN}

\item{files}{list of \code{\link{FileCreateOrUpdateRequest}}}
}
\description{
DatasetPatchRequest per https://api.data.world/v0/swagger.json
}
\examples{
request <- data.world::DatasetPatchRequest(description = "description", summary = "summary",
   tags = list("sdk"), license = "Public Domain")
}
