% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetPutRequest.R
\name{DatasetPutRequest}
\alias{DatasetPutRequest}
\title{DatasetPutRequest per https://api.data.world/v0/swagger.json}
\usage{
DatasetPutRequest(description = NULL, summary = NULL, tags = NULL,
  license = NULL, visibility, files = NULL)
}
\arguments{
\item{description}{Short description for dataset.}

\item{summary}{Dataset summary, markdown supported}

\item{tags}{Array of strings to tag datase}

\item{license}{Public Domain, PDDL, CC-0, CC-BY, ODC-BY, CC-BY-SA, ODC-ODbL, CC BY-NC-SA}

\item{visibility}{PRIVATE or OPEN (required)}

\item{files}{list of \code{\link{FileCreateRequest}}}
}
\description{
DatasetPutRequest per https://api.data.world/v0/swagger.json
}
\examples{
datasetPutRequest <- data.world::DatasetPutRequest(visibility = "OPEN",
 description = "updated description")
}
