% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetSummaryResponse.R
\name{DatasetSummaryResponse}
\alias{DatasetSummaryResponse}
\title{DatasetSummaryResponse per https://api.data.world/v0/swagger.json}
\usage{
DatasetSummaryResponse(structure)
}
\arguments{
\item{structure}{the structured response}
}
\description{
DatasetSummaryResponse per https://api.data.world/v0/swagger.json
}
\seealso{
\code{\link{getDataset}}
}
