% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTypeUtil.R
\name{datatypes}
\alias{datatypes}
\title{Look up types support across platform}
\usage{
datatypes(r = NULL, datapackage = NULL, sparql = NULL, ...)
}
\arguments{
\item{r}{a valid R type (optional)}

\item{datapackage}{a valid datapackage type per http://frictionlessdata.io/data-packages/ (optional)}

\item{sparql}{a valid sparql type per https://www.w3.org/TR/rdf-sparql-query/#abbrevRdfType (optional)}

\item{...}{Further argument , potentially use for future platform}
}
\value{
given a type in one platform , this function return the corresponding types from other
 supported platform
}
\description{
Look up types support across platform
}
\examples{
 allType <- data.world::datatypes()
}
