% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFileAndSyncSources.R
\name{deleteFileAndSyncSources.ApiClient}
\alias{deleteFileAndSyncSources.ApiClient}
\title{delete a file from the dataset}
\usage{
\method{deleteFileAndSyncSources}{ApiClient}(apiClient, dataset, name)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{name}{the filename including the file extension. If a file by that name already
exists in the dataset, the file will be updated/overwritten.}
}
\description{
delete a file from the dataset
}
\examples{
data.world::deleteFileAndSyncSources(data.world()$apiClient, dataset = "ownerid/datasetid",
 name = "file.csv")
}
