% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFileAndSyncSources.R
\name{deleteFilesAndSyncSources.ApiClient}
\alias{deleteFilesAndSyncSources.ApiClient}
\title{delete a file from the dataset}
\usage{
\method{deleteFilesAndSyncSources}{ApiClient}(apiClient, dataset, names)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{names}{list names}
}
\description{
delete a file from the dataset
}
\examples{
data.world::deleteFilesAndSyncSources(data.world()$apiClient, dataset = "ownerid/datasetid",
names = list("file1.csv", "file2.csv"))
}
