% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFileAndSyncSources.R
\name{deleteFilesAndSyncSources}
\alias{deleteFilesAndSyncSources}
\title{delete a file from the dataset}
\usage{
deleteFilesAndSyncSources(connection, dataset, names)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{names}{list names}
}
\description{
delete a file from the dataset
}
\examples{
connection <- data.world(token = "YOUR_API_TOKEN_HERE")
data.world::deleteFilesAndSyncSources(connection, dataset = "ownerid/datasetid", names = list("file1.csv", "file2.csv"))
}
