% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadDatapackage.R
\name{downloadDatapackage.ApiClient}
\alias{downloadDatapackage.ApiClient}
\title{return a data package of the latest dataset version.}
\usage{
\method{downloadDatapackage}{ApiClient}(apiClient, datasetKey)
}
\arguments{
\item{apiClient}{a data.world apiClient}

\item{datasetKey}{a data.world dataset url e.g https://data.world/jonloyens/an-intro-to-dataworld-dataset}

\item{output}{the cache dir where we are storing the datapackage content}
}
\description{
A data package is a standardized way
 to package data for simple sharing between tools and people. Learn more about
 the Data Package specification at frictionlessdata.io.
}
\examples{
\dontrun{
data.world::downloadDatapackage(data.world()$apiClient , "https://data.world/ownerid/datasetid")
}
}
