% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatapackage.R
\name{downloadDatapackage}
\alias{downloadDatapackage}
\title{return a data package of the latest dataset version.}
\usage{
downloadDatapackage(connection, dataset, workingDirId = generateTimestamp(),
  output = sprintf("tmp/dp/\%s/\%s", dataset, workingDirId))
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the  "agentid/datasetid" for the dataset from
which to download the datapackage}

\item{output}{the cache dir where we are storing the datapackage content}
}
\description{
A data package is a standardized way
 to package data for simple sharing between tools and people. Learn more about
 the Data Package specification at frictionlessdata.io.
}
\examples{
\dontrun{
data.world::downloadDatapackage(connection = data.world() , dataset = "ownerid/datasetid")
}
}
