% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile.ApiClient}
\alias{downloadFile.ApiClient}
\title{Download file from the latest dataset version}
\usage{
\method{downloadFile}{ApiClient}(apiClient, dataset, fileName, output)
}
\arguments{
\item{apiClient}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against
which to execute the query}

\item{fileName}{the filename as listed in the data.world dataset}

\item{output}{the local file to store the downloaded content}
}
\description{
Download file from the latest dataset version
}
\examples{
downloadFile(data.world()$apiClient, dataset="ownerid/datasetid" , fileName = "file.csv",
output = "tmp/file.csv")
}
