% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile}
\alias{downloadFile}
\title{Download file from the latest dataset version}
\usage{
downloadFile(connection, dataset, fileName, output)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{fileName}{the filename as listed in the data.world dataset}

\item{output}{the local file to store the downloaded content}
}
\description{
Download file from the latest dataset version
}
