% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFileAsDataFrame.R
\name{downloadFileAsDataFrame.ApiClient}
\alias{downloadFileAsDataFrame.ApiClient}
\title{Download the input file into an R data frame}
\usage{
\method{downloadFileAsDataFrame}{ApiClient}(apiClient, dataset, fileName)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{fileName}{the filename as listed in the data.world dataset
(only csv file is supported for now)}
}
\description{
Download the input file into an R data frame
}
