% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMessage.R
\name{ErrorMessage}
\alias{ErrorMessage}
\title{De-serialize a structured list into ErrorMessage object}
\usage{
ErrorMessage(structure)
}
\arguments{
\item{structure}{structured response}
}
\description{
De-serialize a structured list into ErrorMessage object
}
