% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileBatchUpdateRequest.R
\name{FileBatchUpdateRequest}
\alias{FileBatchUpdateRequest}
\title{FileBatchUpdateRequest per https://api.data.world/v0/swagger.json}
\usage{
FileBatchUpdateRequest(fileSources = list())
}
\arguments{
\item{fileSources}{list of files
request <- data.world::FileBatchUpdateRequest()
request <- data.world::FileBatchUpdateRequest(list(data.world::FileCreateOrUpdateRequest(name, url)))
request <- data.world::addFile(request = request, name = "file.csv", url = "http://data.world/file.csv")}
}
\description{
FileBatchUpdateRequest per https://api.data.world/v0/swagger.json
}
