% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileCreateOrUpdateRequest.R
\name{FileCreateOrUpdateRequest}
\alias{FileCreateOrUpdateRequest}
\title{FileCreateOrUpdateRequest per https://api.data.world/v0/swagger.json}
\usage{
FileCreateOrUpdateRequest(name, url)
}
\arguments{
\item{name}{the filename including the file extension. If a file by that name already exists in the dataset, the file will be updated/overwritten.}

\item{url}{the public, full direct-download path to file}
}
\description{
FileCreateOrUpdateRequest per https://api.data.world/v0/swagger.json
}
\examples{
fileCreateOrUpdateRequest <- data.world::FileCreateOrUpdateRequest(name = "file.csv", url = "https://data.world/file.csv")
}
