% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileCreateOrUpdateRequest.R
\name{FileCreateRequest}
\alias{FileCreateRequest}
\title{FileCreateRequest per https://api.data.world/v0/swagger.json}
\usage{
FileCreateRequest(name, url)
}
\arguments{
\item{name}{the filename including the file extension. If a file by that name already
exists in the dataset, the file will be updated/overwritten.}

\item{url}{the public, full direct-download path to file}
}
\description{
FileCreateRequest per https://api.data.world/v0/swagger.json
}
\examples{
fileCreateRequest <- data.world::FileCreateRequest(name = "file.csv",
 url = "https://data.world/file.csv")
}
