% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceCreateRequest.R
\name{FileSourceCreateOrUpdateRequest}
\alias{FileSourceCreateOrUpdateRequest}
\title{FileSourceCreateOrUpdateRequest per https://api.data.world/v0/swagger.json}
\usage{
FileSourceCreateOrUpdateRequest(url)
}
\arguments{
\item{url}{the public, full direct-download path to file}
}
\description{
FileSourceCreateOrUpdateRequest per https://api.data.world/v0/swagger.json
}
\examples{
fileSourceCreateOrUpdateRequest <- data.world::FileSourceCreateOrUpdateRequest(
"https://open.obamawhitehouse/file3.csv")
}
