% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceCreateRequest.R
\name{FileSourceCreateRequest}
\alias{FileSourceCreateRequest}
\title{FileSourceCreateRequest per https://api.data.world/v0/swagger.json}
\usage{
FileSourceCreateRequest(url)
}
\arguments{
\item{url}{the public, full direct-download path to file}
}
\description{
FileSourceCreateRequest per https://api.data.world/v0/swagger.json
}
\examples{
fileSourceCreateRequest <- data.world::FileSourceCreateRequest(
"https://open.obamawhitehouse/file3.csv")
}
