% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceSummaryResponse.R
\name{FileSourceSummaryResponse}
\alias{FileSourceSummaryResponse}
\title{De-serialize a structured list into FileSourceSummaryResponse object}
\usage{
FileSourceSummaryResponse(structure)
}
\arguments{
\item{structure}{structured response}
}
\description{
De-serialize a structured list into FileSourceSummaryResponse object
}
