% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSummaryResponse.R
\name{FileSummaryResponse}
\alias{FileSummaryResponse}
\title{De-serialize a structured list into FileSummaryResponse object}
\usage{
FileSummaryResponse(structure)
}
\arguments{
\item{structure}{structured response}
}
\description{
De-serialize a structured list into FileSummaryResponse object
}
