% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset.ApiClient}
\alias{getDataset.ApiClient}
\title{Retrieve a dataset via data.world public api}
\usage{
\method{getDataset}{ApiClient}(apiClient, dataset)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\value{
the data.world dataset's metadata
}
\description{
Retrieve a dataset via data.world public api
}
\examples{
getDataset(data.world()$apiClient, dataset="user/dataset")
}
\seealso{
https://docs.data.world/documentation/api/
}
