% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{Retrieve a dataset via data.world public api}
\usage{
getDataset(connection, dataset)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\value{
the data.world dataset's metadata
}
\description{
Retrieve a dataset via data.world public api
}
\examples{
connection <- data.world(token = "YOUR_API_TOKEN_HERE")
getDataset(connection, dataset="user/dataset")
}
\seealso{
https://docs.data.world/documentation/api/
}
