% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-addin.R
\name{insight_project_filter}
\alias{insight_project_filter}
\title{Filter the specified list of projects to those suitable for selection in
the add-in}
\usage{
insight_project_filter(project_list)
}
\arguments{
\item{project_list}{the list of projects}
}
\value{
the list filtered for those suitable for selection
}
\description{
Filter the specified list of projects to those suitable for selection in
the add-in
}
\keyword{internal}
