% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.world.R
\name{loadDataset.data.world}
\alias{loadDataset.data.world}
\title{Declare and load a dataset dependency}
\usage{
\method{loadDataset}{data.world}(datadotworld, datasetKey)
}
\arguments{
\item{datadotworld}{a data.world sdk client}

\item{datasetKey}{a data.world dataset url e.g https://data.world/jonloyens/an-intro-to-dataworld-dataset}
}
\description{
Declare and load a dataset dependency
}
\examples{
\dontrun{
data.world::loadDataset(data.world(), 'https://data.world/jonloyens/an-intro-to-dataworld-dataset')
}
}
