% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{loadSchema}
\alias{loadSchema}
\title{return a schema structure of the table}
\usage{
loadSchema(datapackage, table)
}
\arguments{
\item{datapackage}{a data package}

\item{table}{a valid table name}
}
\description{
return a schema structure of the table
}
