% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDataset.R
\name{LocalDataset}
\alias{LocalDataset}
\title{An adapter class to data package structure generated by https://github.com/frictionlessdata/datapackage-r}
\usage{
LocalDataset(dataset, datapackagePath)
}
\arguments{
\item{dataset}{the  "agentid/datasetid" for the dataset from
which to download the datapackage}

\item{datapackagePath}{path to where datapackage.json is}

\item{structure}{- a structure object returned `datapkg::datapkg_read`}
}
\description{
An adapter class to data package structure generated by https://github.com/frictionlessdata/datapackage-r
}
