% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDataset.R
\name{patchDataset.ApiClient}
\alias{patchDataset.ApiClient}
\title{Update an existing dataset.}
\usage{
\method{patchDataset}{ApiClient}(apiClient, datasetPatchRequest, datasetid)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{datasetPatchRequest}{a \code{\link{DatasetPatchRequest}}}

\item{datasetid}{the "agentid/datasetid" for the dataset against
which to execute the query}
}
\description{
Update an existing dataset.
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")

request <- data.world::DatasetPatchRequest(visibility = "OPEN",
description = "UPDATED DESCRIPTION !")

data.world::patchDataset(data.world()$apiClient, datasetPatchRequest = request,
 datasetid = "agentid/datasetid")
}
