% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{qry_sql}
\alias{qry_sql}
\title{Constructor function for SQL queries.}
\usage{
qry_sql(query_string, params = NULL)
}
\arguments{
\item{query_string}{SQL query string.}

\item{params}{Sequence of positional query parameters.}
}
\value{
Object of type \code{qry_sql}.
}
\description{
Constructor function for SQL queries.
}
\examples{
\dontrun{
  sql_stmt <- data.world::qry_sql("SELECT * FROM Tables")
}
}
