% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.world-deprecated.R
\name{query-deprecated}
\alias{query-deprecated}
\alias{query.data.world}
\title{Execute a SQL or SPARQL query against a data.world client.}
\usage{
\method{query}{data.world}(connection, ...)
}
\description{
Params:
\enumerate{
 \item \code{connection} Object of type data.world.
 \item \code{type} Query type - either "sql" or "sparql".
 \item \code{dataset} Dataset URL or path.
 \item \code{query} SQL or SPARQL query to run.
 \item \code{queryParameters} Query parameters.
}
}
\examples{
connection <- data.world(token = "YOUR_API_TOKEN_HERE")
\dontrun{
query(connection, dataset="user/dataset",
      query="SELECT *
               FROM TableName
              LIMIT 10")

query(connection, dataset="user/dataset",
      query="SELECT *
               FROM TableName where `field1` = ? AND `field2` > ?
              LIMIT 10",
      queryParameters = list("value", 5.0))

query(connection, dataset="user/dataset", type="sparql",
      query="SELECT *
             WHERE {
               ?s ?p ?o.
             } LIMIT 10")

query(connection, dataset="user/dataset", type="sparql",
      query="SELECT *
             WHERE {
             [ :Year ?year ; :Region ?region ; :Indicator_Coverage_and_Disaggregation ?score ]
             FILTER(?score > $v1)
             } LIMIT 10",
             queryParameters = list("$v1"=5.5))
}
}
