% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query.data.world}
\alias{query.data.world}
\title{execute a SQL or SPARQL query against a data.world client}
\usage{
\method{query}{data.world}(datadotworld, type = "sql", datasetKey, query,
  queryParameters = list(), ...)
}
\arguments{
\item{datadotworld}{a data.world sdk cient}

\item{type}{the type of the query - either "sql" or "sparql"}

\item{datasetKey}{a data.world dataset url e.g https://data.world/jonloyens/an-intro-to-dataworld-dataset}

\item{query}{the SQL or SPARQL query to run .}

\item{queryParameters}{Optional comma-separated ?name=value pairs}

\item{...}{additional param}
}
\value{
the query results as a data frame
}
\description{
execute a SQL or SPARQL query against a data.world client
}
\examples{
query(data.world(), "https://data.world/user/dataset",
      query="SELECT *
               FROM TableName
              LIMIT 10")

query(data.world(), "https://data.world/user/dataset",
      query="SELECT *
               FROM TableName where `field1` = ? AND `field2` > ?
              LIMIT 10",
      queryParameters = list("value", 5.0))

query(data.world(), datasetKey = "https://data.world/user/dataset", type="sparql",
      query="SELECT *
             WHERE {
               ?s ?p ?o.
             } LIMIT 10")

query(data.world(), datasetKey = "https://data.world/user/dataset", type="sparql",
      query="SELECT *
             WHERE {
             [ :Year ?year ; :Region ?region ; :Indicator_Coverage_and_Disaggregation ?score ]
             FILTER(?score > $v1)
             } LIMIT 10",
             queryParameters = list("$v1"=5.5))

}
\seealso{
\code{\link{data.world}}
}
