% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDataset.R
\name{replaceDataset.ApiClient}
\alias{replaceDataset.ApiClient}
\title{Replace an existing dataset.}
\usage{
\method{replaceDataset}{ApiClient}(apiClient, datasetPutRequest, datasetid)
}
\arguments{
\item{datasetPutRequest}{a \code{\link{DatasetPutRequest}}}

\item{datasetid}{the "agentid/datasetid" for the dataset against which
to execute the query}

\item{connection}{the connection to data.world}
}
\description{
Replace an existing dataset.
}
\examples{
datasetPutRequest <- data.world::DatasetPutRequest(visibility = "OPEN",
 description = "UPDATED DESCRIPTION !")
data.world::replaceDataset(apiClient = data.world()$apiClient, datasetPutRequest,
 "agentid/datasetid")
}
