% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDataset.R
\name{replaceDataset}
\alias{replaceDataset}
\title{Replace an existing dataset.}
\usage{
replaceDataset(connection, datasetPutRequest, datasetid)
}
\arguments{
\item{connection}{the connection to data.world}

\item{datasetPutRequest}{a \code{\link{DatasetPutRequest}}}

\item{datasetid}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\description{
Replace an existing dataset.
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")
datasetPutRequest <- data.world::DatasetPutRequest(visibility = "OPEN", description = "UPDATED DESCRIPTION !")
data.world::replaceDataset(connection = conn, datasetPutRequest, "agentid/datasetid")
}
