% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_config}
\alias{save_config}
\title{Save configuration to file.}
\usage{
save_config(auth_token, ..., profile = "DEFAULT")
}
\arguments{
\item{auth_token}{API authorization token.}

\item{...}{Reserved for future configuration parameters.}

\item{profile}{Configuration profile.}
}
\value{
Object of type \code{\link{cfg_saved}}
}
\description{
Save configuration to file.
}
\examples{
saved_cfg <- data.world::save_config(auth_token = "YOUR_TOKEN")
data.world::set_config(saved_cfg)
}
