% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-addin.R
\name{save_image_as_insight}
\alias{save_image_as_insight}
\title{Save an image file as a data.world insight}
\usage{
save_image_as_insight(project_id, title, description = NULL, image_file)
}
\arguments{
\item{project_id}{the fully-qualified id of the project to house the insight}

\item{title}{the title of the insight}

\item{description}{the description of the insight (optional)}

\item{image_file}{the file path containing the image}
}
\value{
a list containing the values returned by the upload_file and
create_insight dwapi functions
}
\description{
Save an image file as a data.world insight
}
\keyword{internal}
