% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sdk_version}
\alias{sdk_version}
\title{Return the current data.world version.}
\usage{
sdk_version()
}
\value{
Current package version.
}
\description{
Return the current data.world version.
}
\keyword{internal}
