% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sparql}
\alias{sparql}
\title{Constructor function for SPARQL query statements.}
\usage{
sparql(query, params = NULL)
}
\arguments{
\item{query}{SPARQL query.}

\item{params}{Sequence of named query parameters.}
}
\value{
Object of type \code{sparql}.
}
\description{
Constructor function for SPARQL query statements.
}
