% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sql}
\alias{sql}
\title{Constructor function for SQL query statements.}
\usage{
sql(query, params = NULL)
}
\arguments{
\item{query}{SQL query.}

\item{params}{Sequence of positional query parameters.}
}
\value{
Object of type \code{sql}.
}
\description{
Constructor function for SQL query statements.
}
