% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessMessage.R
\name{SuccessMessage}
\alias{SuccessMessage}
\title{De-serialize a structured list into SuccessMessage object}
\usage{
SuccessMessage(structure)
}
\arguments{
\item{structure}{structured response}
}
\description{
De-serialize a structured list into SuccessMessage object
}
