% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncDataset.R
\name{sync.ApiClient}
\alias{sync.ApiClient}
\title{Fetch latest files from source and update dataset}
\usage{
\method{sync}{ApiClient}(apiClient, dataset)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\description{
Fetch latest files from source and update dataset
}
\examples{
data.world::sync(data.world()$apiClient, dataset="ownerid/dataset")
}
