% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncDataset.R
\name{sync}
\alias{sync}
\title{Fetch latest files from source and update dataset}
\usage{
sync(connection, dataset)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\description{
Fetch latest files from source and update dataset
}
\examples{
connection <- data.world(token = "YOUR_API_TOKEN_HERE")
data.world::sync(connection, dataset="user/dataset")
}
