% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncDataset.R
\name{syncViaGet.ApiClient}
\alias{syncViaGet.ApiClient}
\title{Fetch latest files from source and update dataset via GET, for convenience.}
\usage{
\method{syncViaGet}{ApiClient}(apiClient, dataset)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\description{
Fetch latest files from source and update dataset via GET, for convenience.
}
\examples{
data.world::syncViaGet(data.world()$apiClient, dataset="user/dataset")
}
