% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncDataset.R
\name{syncViaGet}
\alias{syncViaGet}
\title{Fetch latest files from source and update dataset via GET, for convenience.}
\usage{
syncViaGet(connection, dataset)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}
}
\description{
Fetch latest files from source and update dataset via GET, for convenience.
}
\examples{
connection <- data.world(token = "YOUR_API_TOKEN_HERE")
data.world::syncViaGet(connection, dataset="user/dataset")
}
