% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDataFrame.R
\name{uploadDataFrame.ApiClient}
\alias{uploadDataFrame.ApiClient}
\title{Upload a dataframe to a data.world dataset as csv}
\usage{
\method{uploadDataFrame}{ApiClient}(apiClient, dataFrame, fileName, dataset)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataFrame}{the data frame need to be uploaded}

\item{fileName}{the filename (with extension) to be view in the dataset}

\item{dataset}{the data.world datasetid}
}
\description{
Upload a dataframe to a data.world dataset as csv
}
\examples{
df = data.frame(a = c(1,2,3),b = c(4,5,6))
uploadDataFrame(data.world()$apiClient, fileName="sample.csv",
dataFrame = df, dataset = "ownerid/datasetid")
}
