% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDataFrame.R
\name{uploadDataFrame}
\alias{uploadDataFrame}
\title{Upload a dataframe to a data.world dataset as csv}
\usage{
uploadDataFrame(connection, dataFrame, fileName, dataset)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataFrame}{the data frame need to be uploaded}

\item{fileName}{the filename (with extension) to be view in the dataset}

\item{dataset}{the data.world datasetid}
}
\description{
Upload a dataframe to a data.world dataset as csv
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")
df = data.frame(a = c(1,2,3),b = c(4,5,6))
uploadDataFrame(connection = conn, fileName="sample.csv",
dataFrame = df, dataset = "ownerid/datasetid")
}
