% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFile.R
\name{uploadFile.ApiClient}
\alias{uploadFile.ApiClient}
\title{Upload a file to a dataset}
\usage{
\method{uploadFile}{ApiClient}(apiClient, path, fileName, dataset)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{path}{the file path you want to upload}

\item{fileName}{the filename (with extension) to be view in the dataset}

\item{dataset}{datasetid}
}
\value{
http response
}
\description{
Upload a file to a dataset
}
\examples{
df = data.frame(a = c(1,2,3),b = c(4,5,6))
write.csv(df, file = "file.csv")
uploadFile(apiClient = data.world()$apiClient, fileName = "file.csv",
path = "file.csv", dataset = "ownerid/datasetid")
}
