% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFile.R
\name{uploadFile}
\alias{uploadFile}
\title{Upload a file to a dataset}
\usage{
uploadFile(connection, path, fileName, dataset)
}
\arguments{
\item{connection}{the connection to data.world}

\item{path}{the file path you want to upload}

\item{fileName}{the filename (with extension) to be view in the dataset}

\item{dataset}{datasetid}
}
\value{
http response
}
\description{
Upload a file to a dataset
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")
df = data.frame(a = c(1,2,3),b = c(4,5,6))
write.csv(df, file = "file.csv")
uploadFile(connection = conn, "file.csv", "file.csv", "ownerid/datasetid")
}
