% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFile.R
\name{uploadFiles.ApiClient}
\alias{uploadFiles.ApiClient}
\title{Upload one or more files to a dataset.}
\usage{
\method{uploadFiles}{ApiClient}(apiClient, dataset, paths)
}
\arguments{
\item{apiClient}{a data.world api client}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{paths}{the local file paths}
}
\description{
Upload one or more files to a dataset.
}
\examples{
write.csv(data.frame(a = c(1,2,3),b = c(4,5,6)), file = "file1.csv")
write.csv(data.frame(c = c(1,2,3),d = c(4,5,6)), file = "file2.csv")
uploadFiles(apiClient = data.world()$apiClient, dataset = "ownerid/datasetid",
 paths = list ("file1.csv", "file2.csv"))
}
