% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFile.R
\name{uploadFiles}
\alias{uploadFiles}
\title{Upload one or more files to a dataset.}
\usage{
uploadFiles(connection, dataset, paths)
}
\arguments{
\item{connection}{the connection to data.world}

\item{dataset}{the "agentid/datasetid" for the dataset against which to execute the query}

\item{paths}{the local file paths}
}
\description{
Upload one or more files to a dataset.
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")
write.csv(data.frame(a = c(1,2,3),b = c(4,5,6)), file = "file1.csv")
write.csv(data.frame(c = c(1,2,3),d = c(4,5,6)), file = "file2.csv")
uploadFiles(connection = conn, dataset = "ownerid/datasetid",
 paths = list ("file1.csv", "file2.csv"))
}
