% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.world.R
\docType{package}
\name{data.world}
\alias{data.world}
\alias{data.world-package}
\title{data.world: A package for interacting with data.world platform.}
\usage{
data.world(profileName = "default")
}
\description{
data.world: A package for interacting with data.world platform.
}
\section{data.world functions}{

The main SDK expose 2 functions :
\enumerate{
  \item Declare and load a dataset dependency via \code{\link{loadDataset}}
  \item Execute query via \code{\link{query}}
}
}

