% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.world.R
\name{data.world}
\alias{data.world}
\title{data.world client constructor}
\usage{
data.world(token)
}
\arguments{
\item{token}{your data.world API token}
}
\value{
a data.world client
}
\description{
data.world client constructor
}
\seealso{
\code{\link{query}}
}

