% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDataset.R
\name{patchDataset}
\alias{patchDataset}
\title{Update an existing dataset.}
\usage{
patchDataset(connection, datasetPatchRequest, datasetid)
}
\arguments{
\item{connection}{the connection to data.world}

\item{datasetPatchRequest}{a \code{\link{DatasetPatchRequest}}}

\item{datasetid}{the "agentid/datasetid" for the dataset against
which to execute the query}
}
\description{
Update an existing dataset.
}
\examples{
conn <- data.world(token = "YOUR_API_TOKEN_HERE")

request <- data.world::DatasetPatchRequest(visibility = "OPEN",
description = "UPDATED DESCRIPTION !")

data.world::patchDataset(connection = conn, datasetPatchRequest = request,
 datasetid = "agentid/datasetid")
}
