% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{qry_sparql}
\alias{qry_sparql}
\title{Constructor function for SPARQL queries.}
\usage{
qry_sparql(query_string, params = NULL)
}
\arguments{
\item{query_string}{SPARQL query string.}

\item{params}{Sequence of named query parameters.}
}
\value{
Object of type \code{qry_sparql}.
}
\description{
Constructor function for SPARQL queries.
}
\examples{
\dontrun{
  sparql_stmt <- data.world::qry_sparql("SELECT ?s ?p ?o
                                         WHERE {
                                           ?s ?p ?o.
                                         }")
}
}
