% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query.sql}
\alias{query.sparql}
\title{Execute a query on data.world.}
\usage{
query(statement, ...)

\method{query}{sql}(statement, ...)

\method{query}{sparql}(statement, ...)
}
\arguments{
\item{statement}{Query statement of type sql or sparql.}

\item{...}{S3 method specific params.}
}
\value{
Query results as a data frame.
}
\description{
Execute a query on data.world.
}
\section{Methods (by class)}{
\itemize{
\item \code{sql}: Execute a SQL query on data.world.

\item \code{sparql}: Execute a SPARQL query on data.world.
}}

\examples{
\dontrun{
  sql_stmt <- data.world::sql("SELECT * FROM Tables")
  query_results_df <- data.world::query(
    sql_stmt, "jonloyens/an-intro-to-dataworld-dataset")
}
}
\seealso{
\code{\link{sql}} \code{\link{sparql}}
}
